Use possessive quantifiers in regex for decoding cookies
Decoding cookies becomes very complex to handle for the regex engine
when cookie values are very large. This causes `StackOverflowError`s
probable due to the catastrophic backtracking

Use possessive quantifiers to prevent the regex engine from trying
all permutations. This should make the complexity of the regex linear
and not exponential. The performance increase can be significant and
fixes the issue.

Closes #77